+++ shuup/core/models/_suppliers.py
@@ -165,33 +165,6 @@ class Supplier(ModuleInterface, TranslatableShuupModel):
             self.slug = slugify(self.name)
         return super(Supplier, self).save(*args, **kwargs)
 
-    def get_supplier_module_for_product_ids(self, product_ids):
-        supplier_module_dict = {}
-        product_ids = set(product_ids)
-        for index, module in enumerate(self.modules):
-            for product_id in product_ids.copy():
-                if module.can_handle_product(product_id):
-                    if index in supplier_module_dict:
-                        supplier_module_dict[index].append(product_id)
-                    else:
-                        supplier_module_dict[index] = [product_id]
-                    product_ids.remove(product_id)
-        if product_ids:
-            LOGGER.critical(
-                "Products %s did not find a supplier module that can process them."
-                "This most likely is due to that the base supplier module did not load"
-                % (
-                    str(
-                        product_ids,
-                    )
-                )
-            )
-        return supplier_module_dict
-
-    def get_supplier_module_index_for_product_id(self, product_id):
-        supplier_modules = self.get_supplier_module_for_product_ids([product_id])
-        return list(supplier_modules.keys())[0]
-
     def get_orderability_errors(self, shop_product, quantity, customer, *args, **kwargs):
         """
         :param shop_product: Shop Product.
@@ -202,9 +175,14 @@ class Supplier(ModuleInterface, TranslatableShuupModel):
         :type contect: shuup.core.models.Contact
         :rtype: iterable[ValidationError]
         """
-        return self.modules[
-            self.get_supplier_module_index_for_product_id(shop_product.product.id)
-        ].get_orderability_errors(shop_product=shop_product, quantity=quantity, customer=customer, *args, **kwargs)
+        return_dict = {}
+        for module in self.modules:
+            return_dict.update(
+                module.get_orderability_errors(
+                    shop_product=shop_product, quantity=quantity, customer=customer, *args, **kwargs
+                )
+            )
+        return return_dict
 
     def get_stock_statuses(self, product_ids, *args, **kwargs):
         """
@@ -213,8 +191,8 @@ class Supplier(ModuleInterface, TranslatableShuupModel):
         :rtype: dict[int, shuup.core.stocks.ProductStockStatus]
         """
         return_dict = {}
-        for module_index, ids in self.get_supplier_module_for_product_ids(product_ids).items():
-            return_dict.update(self.modules[module_index].get_stock_statuses(ids, *args, **kwargs))
+        for module in self.modules:
+            return_dict.update(module.get_stock_statuses(product_ids, *args, **kwargs))
         return return_dict
 
     def get_stock_status(self, product_id, *args, **kwargs):
@@ -223,9 +201,12 @@ class Supplier(ModuleInterface, TranslatableShuupModel):
         :type product_id: int
         :rtype: shuup.core.stocks.ProductStockStatus
         """
-        return self.modules[self.get_supplier_module_index_for_product_id(product_id)].get_stock_status(
-            product_id, *args, **kwargs
-        )
+        return_object = {}
+        for module in self.modules:
+            return_object = module.get_stock_status(product_id, *args, **kwargs)
+            if return_object:
+                break
+        return return_object
 
     def get_suppliable_products(self, shop, customer):
         """
@@ -245,27 +226,28 @@ class Supplier(ModuleInterface, TranslatableShuupModel):
         from shuup.core.suppliers.base import StockAdjustmentType
 
         adjustment_type = type or StockAdjustmentType.INVENTORY
-        return self.modules[self.get_supplier_module_index_for_product_id(product_id)].adjust_stock(
-            product_id, delta, created_by=created_by, type=adjustment_type, *args, **kwargs
-        )
+        return_object = None
+        for module in self.modules:
+            return_object = module.adjust_stock(
+                product_id, delta, created_by=created_by, type=adjustment_type, *args, **kwargs
+            )
+            if return_object:
+                break
+        return return_object
 
     def update_stock(self, product_id, *args, **kwargs):
-        return self.modules[self.get_supplier_module_index_for_product_id(product_id)].update_stock(
-            product_id, *args, **kwargs
-        )
+        return_dict = {}
+        for module in self.modules:
+            return_dict.update(module.update_stock(product_id, *args, **kwargs))
+        return return_dict
 
     def update_stocks(self, product_ids, *args, **kwargs):
-        for module_index, ids in self.get_supplier_module_for_product_ids(product_ids).items():
-            self.modules[module_index].update_stocks(ids, *args, **kwargs)
+        for module in self.modules:
+            module.update_stocks(product_ids, *args, **kwargs)
 
     def ship_products(self, shipment, product_quantities, *args, **kwargs):
-        product_ids = [product.id for product in product_quantities.keys()]
-        for module_index, ids in self.get_supplier_module_for_product_ids(product_ids).items():
-            product_quantitie = {}
-            for product in product_quantities.keys():
-                if product.id in ids:
-                    product_quantitie[product] = product_quantities[product]
-            self.modules[module_index].ship_products(shipment, product_quantitie, *args, **kwargs)
+        for module in self.modules:
+            module.ship_products(shipment, product_quantities, *args, **kwargs)
 
     def soft_delete(self):
         if not self.deleted:
@@ -290,11 +272,15 @@ class SupplierModule(models.Model):
     module_identifier = models.CharField(
         max_length=64,
         verbose_name=_("module identifier"),
-        help_text=_("Supplier modules define the rules by which inventory is managed."),
+        unique=True,
+        help_text=_(
+            "Select the types of products this supplier can handle."
+            "Example for normal products select just Simple Supplier."
+        ),
     )
     name = models.CharField(
         max_length=64,
-        verbose_name=_("module name"),
+        verbose_name=_("Module name"),
         help_text=_("Supplier modules name."),
     )
 
@@ -310,6 +296,8 @@ class SupplierModule(models.Model):
         for module in get_provide_objects(Supplier.module_provides_key):
             try:
                 cls.objects.get_or_create(module_identifier=module.identifier, defaults={"name": module.name})
+            # Errors that will be raised if this model is change and not migrated yet
+            # As this method will be runned before migrations can be applied we need to catch these errors
             except ProgrammingError:
                 break
             except OperationalError: